requirements.txt
discord.py
python-dotenv
semidbm
requests
